Ref: moving extractObjectSchema() to OpenAPI helpers.
